+++ changelog.md
@@ -26,6 +26,7 @@ Codion Change Log
 - EntityApplicationPanel.createHelpMenuControls() bug fixed, trailing separator removed.
 - EntityDialogs.AddEntityDialogBuilder, EditEntityDialogBuilder.confirm() added, default false.
 - EntityComboBox, EntityComboBoxPanel.Builder, EntitySearchField, EntitySearchFieldPanel.Builder.confirmAdd() and confirmEdit() added.
+- EntityEditComponentPanel.onException() bug fixed, no longer displays cancel exceptions.
 
 ## 0.18.0
 ### is.codion